split burgerItemBuilder from Burger
